requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib[bcrypt]
sqlalchemy
pillow
pytesseract
pdf2image
huggingface-hub
